requirements.txt
torch==2.1.0
torchvision==0.16.0
streamlit==1.28.0
Pillow==10.0.1
numpy==1.24.3